backend/requirements.txt
Django>=4.2.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
